Minor refactoring Moved the buttons and pistons files to their own folder to keep things organized. Also made some of the names lowercase because style.
